+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `@lumx/core`:
     -   Moved `RawClickable` from `@lumx/react`
+    -   Moved `ButtonRoot` from `@lumx/react`
 
 ## [4.1.0][] - 2026-01-22